Replace some uses of off_t by uint64_t.
This allows code that does not rely on lseek etc. to work
even on systems that do not support 64 bit off_t yet.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35975 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ stream/stream_dvdnav.c
@@ -87,7 +87,7 @@ static const struct m_struct_st stream_opts = {
   stream_opts_fields
 };
 
-static int seek(stream_t *s, off_t newpos);
+static int seek(stream_t *s, uint64_t newpos);
 static void show_audio_subs_languages(dvdnav_t *nav);
 
 static dvdnav_priv_t * new_dvdnav_stream(char * filename) {
@@ -276,7 +276,7 @@ static void update_title_len(stream_t *stream) {
 
   status = dvdnav_get_position(priv->dvdnav, &pos, &len);
   if(status == DVDNAV_STATUS_OK && len) {
-    stream->end_pos = (off_t) len * 2048;
+    stream->end_pos = len * 2048ull;
     stream->seek = seek;
   } else {
     stream->seek = NULL;
@@ -285,7 +285,7 @@ static void update_title_len(stream_t *stream) {
 }
 
 
-static int seek(stream_t *s, off_t newpos) {
+static int seek(stream_t *s, uint64_t newpos) {
   uint32_t sector = 0;
   dvdnav_priv_t *priv = s->priv;
