+++ stream/stream_netstream.c
@@ -188,8 +188,8 @@ static int fill_buffer(stream_t *s, char* buffer, int max_len){
 }
 
 
-static int seek(stream_t *s,off_t newpos) {
-  uint64_t pos = le2me_64((uint64_t)newpos);
+static int seek(stream_t *s, uint64_t newpos) {
+  uint64_t pos = le2me_64(newpos);
   mp_net_stream_packet_t* pack;
 
   pack = send_net_stream_cmd(s,NET_STREAM_SEEK,(char*)&pos,8);
